+++ src/components/chip/index.ts
@@ -1,0 +1,1 @@
+export { Chip } from './Chip';
